device implemented

+++ src/main/java/kennesawstate/cs4850/tallulah/domain/Channel.java
@@ -8,6 +8,6 @@ public class Channel {
     String channelName;
     String title;
     String text;
-    double refreshTime;
+    Long refreshTime;
     ChannelType channelType;
 }
